requirements.txt
Django
requests
requests[socks]
dj-database-url
mysqlclient
whitenoise
django-stronghold==0.4.0
djangorestframework==3.11.0
django-configurations==2.2
django-extensions
django-debug-toolbar
django-bootstrap-form==3.4
django-allauth==0.41.0

jsonfield
tqdm

gunicorn

# zbmath
sqlalchemy
psycopg2
pandas
